+++ sycl/include/sycl/detail/cg.hpp
@@ -10,7 +10,6 @@
 
 #include <sycl/accessor.hpp>
 #include <sycl/backend_types.hpp>
-#include <sycl/detail/accessor_impl.hpp>
 #include <sycl/detail/cg_types.hpp>
 #include <sycl/detail/common.hpp>
 #include <sycl/detail/export.hpp>
@@ -170,7 +169,7 @@ public:
   CG(CGTYPE Type, std::vector<std::vector<char>> ArgsStorage,
      std::vector<detail::AccessorImplPtr> AccStorage,
      std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-     std::vector<Requirement *> Requirements,
+     std::vector<AccessorImplHost *> Requirements,
      std::vector<detail::EventImplPtr> Events, detail::code_location loc = {})
       : MType(Type), MArgsStorage(std::move(ArgsStorage)),
         MAccStorage(std::move(AccStorage)),
@@ -221,7 +220,7 @@ private:
 public:
   /// List of requirements that specify which memory is needed for the command
   /// group to be executed.
-  std::vector<Requirement *> MRequirements;
+  std::vector<AccessorImplHost *> MRequirements;
   /// List of events that order the execution of this CG
   std::vector<detail::EventImplPtr> MEvents;
   // Member variables to capture the user code-location
@@ -252,7 +251,7 @@ public:
                std::vector<std::vector<char>> ArgsStorage,
                std::vector<detail::AccessorImplPtr> AccStorage,
                std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-               std::vector<Requirement *> Requirements,
+               std::vector<AccessorImplHost *> Requirements,
                std::vector<detail::EventImplPtr> Events,
                std::vector<ArgDesc> Args, std::string KernelName,
                detail::OSModuleHandle OSModuleHandle,
@@ -303,7 +302,7 @@ public:
          std::vector<std::vector<char>> ArgsStorage,
          std::vector<detail::AccessorImplPtr> AccStorage,
          std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-         std::vector<Requirement *> Requirements,
+         std::vector<AccessorImplHost *> Requirements,
          std::vector<detail::EventImplPtr> Events,
          detail::code_location loc = {})
       : CG(CopyType, std::move(ArgsStorage), std::move(AccStorage),
@@ -318,39 +317,39 @@ public:
 class CGFill : public CG {
 public:
   std::vector<char> MPattern;
-  Requirement *MPtr;
+  AccessorImplHost *MPtr;
 
   CGFill(std::vector<char> Pattern, void *Ptr,
          std::vector<std::vector<char>> ArgsStorage,
          std::vector<detail::AccessorImplPtr> AccStorage,
          std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-         std::vector<Requirement *> Requirements,
+         std::vector<AccessorImplHost *> Requirements,
          std::vector<detail::EventImplPtr> Events,
          detail::code_location loc = {})
       : CG(Fill, std::move(ArgsStorage), std::move(AccStorage),
            std::move(SharedPtrStorage), std::move(Requirements),
            std::move(Events), std::move(loc)),
-        MPattern(std::move(Pattern)), MPtr((Requirement *)Ptr) {}
-  Requirement *getReqToFill() { return MPtr; }
+        MPattern(std::move(Pattern)), MPtr((AccessorImplHost *)Ptr) {}
+  AccessorImplHost *getReqToFill() { return MPtr; }
 };
 
 /// "Update host" command group class.
 class CGUpdateHost : public CG {
-  Requirement *MPtr;
+  AccessorImplHost *MPtr;
 
 public:
   CGUpdateHost(void *Ptr, std::vector<std::vector<char>> ArgsStorage,
                std::vector<detail::AccessorImplPtr> AccStorage,
                std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-               std::vector<Requirement *> Requirements,
+               std::vector<AccessorImplHost *> Requirements,
                std::vector<detail::EventImplPtr> Events,
                detail::code_location loc = {})
       : CG(UpdateHost, std::move(ArgsStorage), std::move(AccStorage),
            std::move(SharedPtrStorage), std::move(Requirements),
            std::move(Events), std::move(loc)),
-        MPtr((Requirement *)Ptr) {}
+        MPtr((AccessorImplHost *)Ptr) {}
 
-  Requirement *getReqToUpdate() { return MPtr; }
+  AccessorImplHost *getReqToUpdate() { return MPtr; }
 };
 
 /// "Copy USM" command group class.
@@ -364,7 +363,7 @@ public:
             std::vector<std::vector<char>> ArgsStorage,
             std::vector<detail::AccessorImplPtr> AccStorage,
             std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-            std::vector<Requirement *> Requirements,
+            std::vector<AccessorImplHost *> Requirements,
             std::vector<detail::EventImplPtr> Events,
             detail::code_location loc = {})
       : CG(CopyUSM, std::move(ArgsStorage), std::move(AccStorage),
@@ -388,7 +387,7 @@ public:
             std::vector<std::vector<char>> ArgsStorage,
             std::vector<detail::AccessorImplPtr> AccStorage,
             std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-            std::vector<Requirement *> Requirements,
+            std::vector<AccessorImplHost *> Requirements,
             std::vector<detail::EventImplPtr> Events,
             detail::code_location loc = {})
       : CG(FillUSM, std::move(ArgsStorage), std::move(AccStorage),
@@ -410,7 +409,7 @@ public:
                 std::vector<std::vector<char>> ArgsStorage,
                 std::vector<detail::AccessorImplPtr> AccStorage,
                 std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-                std::vector<Requirement *> Requirements,
+                std::vector<AccessorImplHost *> Requirements,
                 std::vector<detail::EventImplPtr> Events,
                 detail::code_location loc = {})
       : CG(PrefetchUSM, std::move(ArgsStorage), std::move(AccStorage),
@@ -432,7 +431,7 @@ public:
               std::vector<std::vector<char>> ArgsStorage,
               std::vector<detail::AccessorImplPtr> AccStorage,
               std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-              std::vector<Requirement *> Requirements,
+              std::vector<AccessorImplHost *> Requirements,
               std::vector<detail::EventImplPtr> Events, CGTYPE Type,
               detail::code_location loc = {})
       : CG(Type, std::move(ArgsStorage), std::move(AccStorage),
@@ -452,7 +451,7 @@ public:
                 std::vector<std::vector<char>> ArgsStorage,
                 std::vector<detail::AccessorImplPtr> AccStorage,
                 std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-                std::vector<Requirement *> Requirements,
+                std::vector<AccessorImplHost *> Requirements,
                 std::vector<detail::EventImplPtr> Events, CGTYPE Type,
                 detail::code_location loc = {})
       : CG(Type, std::move(ArgsStorage), std::move(AccStorage),
@@ -477,7 +476,7 @@ public:
              std::vector<std::vector<char>> ArgsStorage,
              std::vector<detail::AccessorImplPtr> AccStorage,
              std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-             std::vector<Requirement *> Requirements,
+             std::vector<AccessorImplHost *> Requirements,
              std::vector<detail::EventImplPtr> Events, CGTYPE Type,
              detail::code_location loc = {})
       : CG(Type, std::move(ArgsStorage), std::move(AccStorage),
@@ -495,7 +494,7 @@ public:
             std::vector<std::vector<char>> ArgsStorage,
             std::vector<detail::AccessorImplPtr> AccStorage,
             std::vector<std::shared_ptr<const void>> SharedPtrStorage,
-            std::vector<Requirement *> Requirements,
+            std::vector<AccessorImplHost *> Requirements,
             std::vector<detail::EventImplPtr> Events, CGTYPE Type,
             detail::code_location loc = {})
       : CG(Type, std::move(ArgsStorage), std::move(AccStorage),
