+++ sycl/source/accessor.cpp
@@ -15,6 +15,65 @@ namespace detail {
 device getDeviceFromHandler(handler &CommandGroupHandlerRef) {
   return CommandGroupHandlerRef.MQueue->get_device();
 }
+
+AccessorBaseHost::AccessorBaseHost(id<3> Offset, range<3> AccessRange,
+                                   range<3> MemoryRange,
+                                   access::mode AccessMode, void *SYCLMemObject,
+                                   int Dims, int ElemSize, int OffsetInBytes,
+                                   bool IsSubBuffer) {
+  impl = std::shared_ptr<AccessorImplHost>(
+      new AccessorImplHost(Offset, AccessRange, MemoryRange, AccessMode,
+                           (detail::SYCLMemObjI *)SYCLMemObject, Dims, ElemSize,
+                           OffsetInBytes, IsSubBuffer));
+}
+
+id<3> &AccessorBaseHost::getOffset() { return impl->MOffset; }
+range<3> &AccessorBaseHost::getAccessRange() { return impl->MAccessRange; }
+range<3> &AccessorBaseHost::getMemoryRange() { return impl->MMemoryRange; }
+void *AccessorBaseHost::getPtr() { return impl->MData; }
+unsigned int AccessorBaseHost::getElemSize() const { return impl->MElemSize; }
+
+const id<3> &AccessorBaseHost::getOffset() const { return impl->MOffset; }
+const range<3> &AccessorBaseHost::getAccessRange() const {
+  return impl->MAccessRange;
+}
+const range<3> &AccessorBaseHost::getMemoryRange() const {
+  return impl->MMemoryRange;
+}
+void *AccessorBaseHost::getPtr() const {
+  return const_cast<void *>(impl->MData);
+}
+
+void *AccessorBaseHost::getMemoryObject() const { return impl->MSYCLMemObj; }
+
+LocalAccessorBaseHost::LocalAccessorBaseHost(sycl::range<3> Size, int Dims,
+                                             int ElemSize) {
+  impl = std::shared_ptr<LocalAccessorImplHost>(
+      new LocalAccessorImplHost(Size, Dims, ElemSize));
+}
+sycl::range<3> &LocalAccessorBaseHost::getSize() { return impl->MSize; }
+const sycl::range<3> &LocalAccessorBaseHost::getSize() const {
+  return impl->MSize;
+}
+void *LocalAccessorBaseHost::getPtr() {
+  // Const cast this in order to call the const getPtr.
+  return const_cast<const LocalAccessorBaseHost *>(this)->getPtr();
+}
+void *LocalAccessorBaseHost::getPtr() const {
+  char *ptr = impl->MMem.data();
+
+  // Align the pointer to MElemSize.
+  size_t val = reinterpret_cast<size_t>(ptr);
+  if (val % impl->MElemSize != 0) {
+    ptr += impl->MElemSize - val % impl->MElemSize;
+  }
+
+  return ptr;
+}
+
+int LocalAccessorBaseHost::getNumOfDims() { return impl->MDims; }
+int LocalAccessorBaseHost::getElementSize() { return impl->MElemSize; }
+
 } // namespace detail
 } // __SYCL_INLINE_VER_NAMESPACE(_V1)
 } // namespace sycl
