+++ sycl/source/detail/scheduler/commands.hpp
@@ -17,10 +17,10 @@
 #include <unordered_set>
 #include <vector>
 
+#include <detail/accessor_impl.hpp>
 #include <detail/event_impl.hpp>
 #include <detail/program_manager/program_manager.hpp>
 #include <sycl/access/access.hpp>
-#include <sycl/detail/accessor_impl.hpp>
 #include <sycl/detail/cg.hpp>
 
 namespace sycl {
