+++ sycl/source/interop_handle.cpp
@@ -6,11 +6,11 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include <detail/accessor_impl.hpp>
 #include <detail/backend_impl.hpp>
 #include <detail/context_impl.hpp>
 #include <detail/device_impl.hpp>
 #include <detail/queue_impl.hpp>
-#include <sycl/detail/accessor_impl.hpp>
 #include <sycl/exception.hpp>
 #include <sycl/interop_handle.hpp>
